+++ example/example.dart
@@ -1,12 +1,5 @@
 import 'package:cancelable_compute/cancelable_compute.dart';
 
-Future<void> main() async {
-  final operation = compute(delayed, 5);
-  Future<void>.delayed(Duration(seconds: 1), operation.cancel);
-  final result = await operation.value;
-  print(result ?? -1);
-}
-
 int fib(int n) {
   if (n < 2) {
     return n;
@@ -15,6 +8,15 @@ int fib(int n) {
   return fib(n - 2) + fib(n - 1);
 }
 
-Future<int> delayed(int seconds) {
-  return Future.delayed(Duration(seconds: seconds), () => seconds);
+Future<int> delayed(int n) async {
+  await Future<void>.delayed(Duration(seconds: n ~/ 2));
+  return fib(n);
+}
+
+Future<void> main() async {
+  final operation = compute(delayed, 4);
+  Future<void>.delayed(Duration(seconds: 1), operation.cancel);
+
+  final result = await operation.value;
+  print(result ?? -1);
 }