+++ lib/src/js.dart
@@ -1,53 +1,54 @@
 import 'dart:async' show Completer, FutureOr;
 
-import 'types.dart';
+import 'package:cancelable_compute/src/types.dart';
 
+/// {@macro compute}
 ComputeOperation<R> compute<Q, R>(ComputeCallback<Q, R> callback, Q message) {
   return _Operation<Q, R>(callback, message);
 }
 
-class _Operation<Q, R> implements ComputeOperation<R> {
+final class _Operation<Q, R> implements ComputeOperation<R> {
   _Operation(ComputeCallback<Q, R> callback, Q message)
-      : completer = Completer<R?>(),
-        canceled = false {
+      : _completer = Completer<R?>(),
+        _canceled = false {
     Future<R>(() => callback(message))
         .then<void>(complete)
         .catchError(completeError);
   }
 
-  final Completer<R?> completer;
+  final Completer<R?> _completer;
 
-  bool canceled;
+  bool _canceled;
 
   @override
   bool get isCanceled {
-    return canceled;
+    return _canceled;
   }
 
   bool get isCompleted {
-    return completer.isCompleted;
+    return _completer.isCompleted;
   }
 
   @override
   Future<R?> get value {
-    return completer.future;
+    return _completer.future;
   }
 
   @override
   void cancel([FutureOr<R?>? value]) {
-    canceled = true;
-    completer.complete(value);
+    _canceled = true;
+    _completer.complete(value);
   }
 
   void complete(FutureOr<R?>? data) {
-    if (!canceled) {
-      completer.complete(data);
+    if (!_canceled) {
+      _completer.complete(data);
     }
   }
 
   void completeError(Object error, [StackTrace? stackTrace]) {
-    if (!canceled) {
-      completer.completeError(error, stackTrace);
+    if (!_canceled) {
+      _completer.completeError(error, stackTrace);
     }
   }
 }