+++ lib/src/types.dart
@@ -1,14 +1,16 @@
 import 'dart:async' show FutureOr;
 
+import 'package:cancelable_compute/cancelable_compute.dart';
+
 /// Signature for the callback passed to [compute].
 typedef ComputeCallback<Q, R> = FutureOr<R> Function(Q message);
 
-/// [compute] signature.
+/// [compute] function signature.
 typedef Compute = ComputeOperation<R> Function<Q, R>(
     ComputeCallback<Q, R> callback, Q message);
 
-/// [compute] cancellable operation.
-abstract class ComputeOperation<R> {
+/// [compute] cancellable operation interface.
+abstract interface class ComputeOperation<R> {
   bool get isCanceled;
 
   Future<R?> get value;